refactor(api): split live trade monitoring service into focused services
- Extract alerts, algorithms, comparison, orders, overview, slippage, and user activity into dedicated services
- Move shared helpers into live-trade-monitoring.utils with unit tests
- Slim down live-trade-monitoring.service and update module wiring
- Remove live-trade-monitoring.service entirely and wire controller directly to focused services
- Expand utils spec coverage and refine orders/overview service logic
- Extract calculateDeviationPercent into shared utils, deduplicate from alerts and comparison services
- Use latestPerformanceCondition helper and period filtering in comparison getLiveMetrics query
- Add calculateDeviationPercent unit tests
- Fix user-activity query to include createdAt in select

Closes #334
